+++ test/denormalized_has_one_test.rb
@@ -146,16 +146,9 @@ class DenormalizedHasOneTest < IdentityCache::TestCase
     end
   end
 
-  def test_cache_without_guessable_inverse_name_raises
+  def test_cache_association_known_inverse_raises
     assert_raises IdentityCache::InverseAssociationError do
       Item.cache_has_one(:no_inverse_of_record, embed: true)
-      IdentityCache.eager_load!
-    end
-  end
-
-  def test_cache_without_guessable_inverse_name_does_not_raise_when_inverse_name_specified
-    assert_nothing_raised do
-      Item.cache_has_one(:no_inverse_of_record, inverse_name: :owner, embed: true)
       IdentityCache.eager_load!
     end
   end
@@ -169,7 +162,7 @@ class DenormalizedHasOneTest < IdentityCache::TestCase
 
   def test_cache_has_one_on_derived_model_raises
     assert_raises(IdentityCache::DerivedModelError) do
-      StiRecordTypeA.cache_has_one(:polymorphic_record, inverse_name: :owner, embed: true)
+      StiRecordTypeA.cache_has_one(:polymorphic_record, embed: true)
     end
   end